refactor(tui): extract ansi_parse.rs from tui_render.rs (#472)
Move ANSI→Span parsing + fallback stripper + 5 tests into
dedicated ansi_parse module. Keeps tui_render.rs at 551 lines
(back under the 600-line cap).
